Remove PLATFORM_TYPE required var from oc commatrix command
We originally introduced the PLATFORM_TYPE environment variable to handle cases where the platform type was None. However, None is a valid platform type on its own and should not be interpreted as either a cloud or baremetal platform.

Therefore, we are removing all usage of the PLATFORM_TYPE variable and will rely solely on the existing functions that detect the platform type automatically.