[release-4.18] Remove port 10300 and 10390 from doc commatrix and static entries
Ports 10300, 10390 had been changed to be used only internally and there for shouldn't be documented in the communication matrix.
Note: types.CloudStaticEntriesWorker was deleted as it only contained the above ports.

This is a manual cherry pick of [PR#125](https://github.com/openshift-kni/commatrix/pull/125), as automated cherry pick encountered merge conflicts.