Update test validates the commatrix in OCP docs is up-to-date
The current test is based on the old commatrix docs format the included only one matrix. Hence, we exclude static entries with platform and deployment types, that differ from the ones the cluster the test is running on has, from the documented commatrix when comparing it to the generated one.
The new version (4.18 and above) contains matrices by platform type and deployment, and each cluster is compared to the corresponding matrix. As a results,  the exclusion of the static entries by deployment and platform types from the documented commatrix is not needed.