Improve supported cluster's platform type check
Instead of checking if the cluster's platform type is supported within
the `GetPlatformType` function, we will check and handle the platform
type out side of the function to avoid the assumption the function is
always called in order to get the platform type.

Also if cluster's platform type is not supported, the e2e tests will be
skipped.