[release-4.18] Duplicate bm-sno commatrix under the name none-sno
This PR is the first of changing the bm-sno commmatrix name to none-sno.
after this PR will be merged, the pointer to the commatrix in the
official documentation will be changed, and than we can remove the
commatrix under the name bm-sno.

This PR was manully cherry picked from PR#139