Create a Dockerfile that builds commatrix binary
This PR creates a Dockerfile that builds the commatrix binary in order to
use the commatrix plugin in the origin tests. After the merge of this PR
we will add a prow configuration in the release repo in order to build
the image and promote it using the ci operator, and than use the build
image as a base image for building the origin's repo podman that runs
the origin tests (see
[dockerfile](https://github.com/openshift/origin/blob/main/images/tests/Dockerfile.rhel) so the it will contain the necessary plugin.
The image will also be promoted using ART.
Note: Final image temporarly contains golang, will be removed after the e2e tests removed from our repo.